+++ profiles/latest/databoxedge/mgmt/databoxedge/databoxedgeapi/models.go
@@ -19,20 +19,22 @@
 
 package databoxedgeapi
 
-import original "github.com/Azure/azure-sdk-for-go/services/databoxedge/mgmt/2019-08-01/databoxedge/databoxedgeapi"
+import original "github.com/Azure/azure-sdk-for-go/services/databoxedge/mgmt/2020-12-01/databoxedge/databoxedgeapi"
 
+type AddonsClientAPI = original.AddonsClientAPI
 type AlertsClientAPI = original.AlertsClientAPI
+type AvailableSkusClientAPI = original.AvailableSkusClientAPI
 type BandwidthSchedulesClientAPI = original.BandwidthSchedulesClientAPI
 type ContainersClientAPI = original.ContainersClientAPI
 type DevicesClientAPI = original.DevicesClientAPI
 type JobsClientAPI = original.JobsClientAPI
+type MonitoringConfigClientAPI = original.MonitoringConfigClientAPI
 type NodesClientAPI = original.NodesClientAPI
 type OperationsClientAPI = original.OperationsClientAPI
 type OperationsStatusClientAPI = original.OperationsStatusClientAPI
 type OrdersClientAPI = original.OrdersClientAPI
 type RolesClientAPI = original.RolesClientAPI
 type SharesClientAPI = original.SharesClientAPI
-type SkusClientAPI = original.SkusClientAPI
 type StorageAccountCredentialsClientAPI = original.StorageAccountCredentialsClientAPI
 type StorageAccountsClientAPI = original.StorageAccountsClientAPI
 type TriggersClientAPI = original.TriggersClientAPI